+++ system/kernel/simul_efun.c
@@ -1,7 +1,6 @@
 #include <ansi.h>
 #include "simul_efun/all_environment.c"
 #include "simul_efun/array.c"
-#include "simul_efun/atoi.c"
 #include "simul_efun/bit.c"
 #include "simul_efun/bitmap_font.c"
 #include "simul_efun/break_string.c"